chore: rework reference official component to use react-query
