+++ src/js/references/api.ts
@@ -34,8 +34,16 @@ export function cloneReference({ name, description, refId }): Promise<ReferenceM
         .then(res => res.body);
 }
 
-export function remoteReference({ remote_from }) {
-    return Request.post("/refs").send({ remote_from });
+/**
+ * Remotely installs a reference from the Virtool repository
+ *
+ * @param remote_from - The repository address
+ * @returns A promise resolving to remotely installing a reference
+ */
+export function remoteReference(remote_from: string): Promise<Reference> {
+    return Request.post("/refs")
+        .send({ remote_from })
+        .then(res => res.body);
 }
 
 /**
